Create html for body scan results; Add 3D models to html.
