+++ html/ai-training-female-body-scan-result.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Welcome to FitUp</title>
+    <meta name="Comp2800 template" content="FitUp">
+
+    <!-- Required meta tags -->
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <link rel="stylesheet" type="text/css" href="/styles/body-scan.css" />
+   
+</head>
+
+<body>
+
+    <main>
+        <div id="container3D"></div>
+    </main>
+
+    <script type="module" src="/scripts/ai-training-female-body-scan-result.js"></script> 
+</body>
+
+</html>
