+++ html/ai-training-female-body-scan.html
@@ -11,6 +11,9 @@
 
     <link rel="stylesheet" type="text/css" href="/styles/body-scan.css" />
     <style>
+        body {
+            overflow-y: hidden;
+        }
         video {
             width: 100vw;
             height: 100vh;
@@ -26,7 +29,7 @@
         <div id="container3D"></div>
     </main> -->
     <video loop="" autoplay="" muted="" playsinline="">
-        <source src="/models/female-body-scan.mp4" type="video/mp4">
+        <source src="/models/female-body/female-body-scan.mp4" type="video/mp4">
     </video>   
     <!-- <script type="module" src="/scripts/body-scan.js"></script>  -->
 </body>
