+++ html/ai-training-male-body-scan.html
@@ -11,6 +11,9 @@
 
     <link rel="stylesheet" type="text/css" href="/styles/body-scan.css" />
     <style>
+        body {
+            overflow: hidden;
+        }
         video {
             width: 100vw;
             height: 100vh;
